ref: move collection content to its own scope
