+++ pkg/test.js
@@ -251,6 +251,78 @@ describe("PubkySpecs Example Objects Tests", () => {
         );
       });
     });
+
+    describe("Collection posts", () => {
+      const collectionItemUri = `pubky://${RIO}/pub/pubky.app/posts/0033SREKPC4N0`;
+      const coverImageUrl = `pubky://${RIO}/pub/pubky.app/files/0034A0X7NJ52G`;
+
+      it("should create collection post with JSON envelope content", () => {
+        assert.strictEqual(
+          typeof specsBuilder.createCollectionPost,
+          "function",
+          "PubkySpecsBuilder should expose createCollectionPost"
+        );
+
+        const { post, meta } = specsBuilder.createCollectionPost(
+          "Favorite posts",
+          "Posts worth revisiting",
+          [collectionItemUri],
+          coverImageUrl
+        );
+
+        assert.ok(meta.id, "Collection post should have an ID");
+        assert.ok(meta.url, "Collection post should have a URL");
+        const postChunks = meta.url.split("/");
+        assert.strictEqual(postChunks[2], OTTO, "URL should contain user ID");
+        assert.strictEqual(postChunks[5], "posts", "URL should contain posts path");
+        assert.strictEqual(postChunks[6], meta.id, "URL should contain post ID");
+
+        const postJson = post.toJson();
+        assert.strictEqual(postJson.kind, "collection", "Post kind should be collection");
+        assert.ok(
+          postJson.attachments === null || postJson.attachments === undefined,
+          "Collection items should not be stored in post.attachments"
+        );
+
+        const envelope = JSON.parse(postJson.content);
+        assert.strictEqual(envelope.name, "Favorite posts", "Collection name should match");
+        assert.strictEqual(
+          envelope.description,
+          "Posts worth revisiting",
+          "Collection description should match"
+        );
+        assert.deepStrictEqual(envelope.items, [collectionItemUri], "Collection items should match");
+        assert.strictEqual(envelope.cover_image, coverImageUrl, "Collection cover image should match");
+      });
+
+      it("cannot create collection post with too many items", () => {
+        assert.strictEqual(
+          typeof specsBuilder.createCollectionPost,
+          "function",
+          "PubkySpecsBuilder should expose createCollectionPost"
+        );
+
+        const tooManyItems = Array.from(
+          { length: validationLimits.collectionItemsMaxCount + 1 },
+          (_, index) => `pubky://${RIO}/pub/pubky.app/posts/${String(index).padStart(13, "0")}`
+        );
+
+        assert.throws(
+          () => {
+            specsBuilder.createCollectionPost("Too many", null, tooManyItems, null);
+          },
+          (err) => {
+            const msg = err instanceof Error ? err.message : String(err);
+            assert.ok(
+              msg.includes(`${validationLimits.collectionItemsMaxCount} items`),
+              `Expected collection item limit error, got: "${msg}"`
+            );
+            return true;
+          },
+          "Expected validation error for too many collection items"
+        );
+      });
+    });
   });
 
   describe("Bookmark Pubky-app-specs", () => {
