+++ src/models/post/content/mod.rs
@@ -1,0 +1,3 @@
+pub mod collection;
+
+pub use collection::PubkyAppCollectionContent;
